changed setup for directories and geo placement.

+++ tigre/Utilities/filtering.py
@@ -1,6 +1,6 @@
 from __future__ import print_function
 from __future__ import division
-from Utilities.parkerweight import parkerweight
+from tigre.Utilities.parkerweight import parkerweight
 import numpy as np
 
 import warnings
@@ -21,7 +21,7 @@ def filtering(proj,geo,angles,parker):
     for i in range(len(angles)):
         fproj=np.zeros((filt_len,geo.nDetector[0]),dtype=np.float32)
 
-        fproj[filt_len/2-geo.nDetector[0]/2:filt_len/2+geo.nDetector[0]/2,:]=proj[i]
+        fproj[int(filt_len/2-geo.nDetector[0]/2):int(filt_len/2+geo.nDetector[0]/2),:]=proj[i]
 
         fproj=np.fft.fft(fproj,axis=0)
 
@@ -29,7 +29,7 @@ def filtering(proj,geo,angles,parker):
 
         fproj=np.real(np.fft.ifft(fproj,axis=0))
         end=len(fproj)
-        proj[i]=np.transpose(fproj[end/2-geo.nDetector[0]/2:end/2+geo.nDetector[0]/2,:]/2/geo.dDetector[0]*(2*np.pi/
+        proj[i]=np.transpose(fproj[int(end/2-geo.nDetector[0]/2):int(end/2+geo.nDetector[0]/2),:]/2/geo.dDetector[0]*(2*np.pi/
                                                                                                  len(angles)
                                                                                                  )/2*(geo.DSD/geo.DSO))
 
@@ -38,7 +38,7 @@ def filtering(proj,geo,angles,parker):
 def ramp_flat(n):
     nn=np.arange(-n/2,n/2)
     h=np.zeros(nn.shape,dtype=np.float32)
-    h[n/2]=1/4
+    h[int(n/2)]=1/4
     odd=nn%2==1
     h[odd]=-1/(np.pi*nn[odd])**2
     return h, nn
@@ -46,7 +46,7 @@ def ramp_flat(n):
 def filter(filter,kernel,order,d):
     f_kernel=abs(np.fft.fft(kernel))*2
 
-    filt=f_kernel[:(order/2)+1]
+    filt=f_kernel[:int((order/2)+1)]
     w=2*np.pi*np.arange(len(filt))/order
     if filter not in ['ram_lak','shepp_logan','cosine','hamming','hann',None]:
         raise ValueError('filter not recognised: '+filter)
